requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
bleach==4.1.0
bootstrap4==0.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
crispy-tailwind==0.5.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.11
django-appconf==1.0.5
django-bootstrap-datepicker-plus==4.0.0
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-filter==21.1
django-flatpickr==1.0.1
django-js-asset==2.0.0
django-model-utils==4.2.0
django-notification==1.3.3
django-notifications-hq==1.7.0
django-user-accounts==3.2.0
entrypoints==0.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==7.32.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonfield==3.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mistune==0.8.4
mysqlclient==2.1.0
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
terminado==0.13.3
testpath==0.6.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
Unipath==1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
zipp==3.7.0
